requirements.txt
pyqrcode==1.2.1
Pillow
djangorestframework
pypng
myqr